refactor: update user profile API and update user types to include mutual users
