--- server/api/users/[username]/profile/index.get.ts
@@ -1,14 +1,0 @@
-import { defineWrappedResponseHandler } from '~~/server/utils/handler';
-import usernameParamsSchema from '~~/server/schemas/username';
-
-export default defineWrappedResponseHandler(async (event) => {
-  const { username } = await getValidatedRouterParams(event, (data) =>
-    usernameParamsSchema.validate(data),
-  );
-  const fetcher = serverApiFetch(event);
-
-  const response = await fetcher<ApiSuccessResponse<User>>(`/users/${username}/profile`, {
-    method: 'GET',
-  });
-  return response;
-});
